KIDS_TOOLS/requirements.txt
#요구사항 
Django==3.2.20
djangorestframework==3.14.0
django-extensions==3.2.3
drf-yasg==1.21.7
graphviz==0.20.3
pydotplus==2.0.2